+++ sdks/outpost-typescript/src/lib/config.ts
@@ -73,8 +73,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.1",
-  sdkVersion: "0.3.0",
-  genVersion: "2.669.0",
+  sdkVersion: "0.4.0",
+  genVersion: "2.691.6",
   userAgent:
-    "speakeasy-sdk/typescript 0.3.0 2.669.0 0.0.1 @hookdeck/outpost-sdk",
+    "speakeasy-sdk/typescript 0.4.0 2.691.6 0.0.1 @hookdeck/outpost-sdk",
 } as const;